refactor(fmt): Rename buffer strategy mods
