+++ src/fmt/writer/buffer/plain.rs
