+++ src/fmt/writer/buffer/termcolor.rs
